MicroServices-based-DeepFake-Detection-System-using-Docker/face_detection_service/requirements.txt
fastapi
uvicorn[standard]
face-detection
ffmpeg-python
python-multipart
opencv-python-headless
httpx
torch==2.3.1+cpu
torchvision==0.18.1+cpu
transformers==4.39.3
--extra-index-url https://download.pytorch.org/whl/cpu